Working VisualC suport in IDE's (dll's are still missing). Fixed a few warnings.
git-svn-id: file:///fltk/svn/fltk/branches/branch-1.3@7147 ea41ed52-d2ee-0310-a9c1-e6b18d33e121

+++ fluid/Fl_Function_Type.cxx
@@ -869,7 +869,7 @@ void Fl_Data_Type::write_code1() {
   if (!c) return;
   const char *fn = filename_;
   char *data = 0;
-  size_t nData = -1;
+  int nData = -1;
   // path should be set correctly already
   if (filename_ && !write_sourceview) {
     FILE *f = fopen(filename_, "rb");
